applied pundit to destroy in posts controller, added user_policy called post_owner? and applied pundit on view to remove the delete link when not the owner of the post?

+++ app/controllers/posts_controller.rb
@@ -17,6 +17,7 @@ class PostsController < ApplicationController
 
   def destroy
     @post = Post.find(params[:id])
+    authorize current_user, :owner?, policy_class: UserPolicy
     if @post.destroy
       redirect_to user_posts_path
     else
